Setup Bounty service method that uses Prisma
